Have ForageErrorDetails live in .polling
We faced a tradeoff of prioritizing an organized repo or happy forage
partners. Forage is now at a point where we have to prioritize happy
Forage partners

This PR moves the ForageErrorDetails back to it's, unfortunate, original
home so that we avoid introducing a breaking change while releasing sync
endpoints.
